Added new success messages and renamed services
