+++ microservices/collabService/collabService.Dockerfile
@@ -3,4 +3,4 @@ WORKDIR /app
 COPY . .
 RUN yarn install --production
 CMD ["node", "app.js"]
-EXPOSE 8000
+EXPOSE 2000
