+++ microservices/collabService/package.json
@@ -11,14 +11,8 @@
   "license": "ISC",
   "dependencies": {
     "axios": "^1.5.1",
-    "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
+    "chai-http": "^4.4.0",
     "express": "^4.18.2",
-    "firebase-admin": "^11.11.0",
-    "i": "^0.3.7",
-    "mongodb": "^6.1.0",
-    "mongoose": "^7.5.2",
-    "npm": "^10.2.0",
-    "uuid": "^9.0.1"
+    "socket.io": "^4.7.2"
   }
 }
