+++ microservices/collabService/src/model/RoomModel.js
